# packages/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Hinter Factory ML API
  version: 0.1.0
  description: >-
    Every scoped endpoint (documents, tags, labeling functions, LF runs,
    gold labels, probabilistic labels, evaluation) requires a `project_id`
    query parameter (or, for multipart uploads, a `project_id` form field).
    There is no implicit fallback project; calls that omit it return 400.
    Use `GET /v1/projects` to discover available projects and
    `POST /v1/projects` to create one. The web UI threads `project_id`
    through every request automatically via its project context.
servers:
  - url: http://localhost:8000
...
paths:
...
  /v1/lf-runs:
    get:
      tags: [lfRuns]
      operationId: listLfRuns
      parameters:
        - in: query
          name: tag_id
          schema: { type: string }
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/LfRunStatus"
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LfRun"
    post:
      tags: [lfRuns]
      operationId: createLfRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LfRunCreate"
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LfRun"
  /v1/lf-runs/{run_id}:
    get:
      tags: [lfRuns]
      operationId: getLfRun
      parameters:
        - $ref: "#/components/parameters/LfRunId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LfRun"
  /v1/lf-runs/{run_id}/matrix:
    get:
      tags: [lfRuns]
      operationId: exportLfMatrix
      parameters:
        - $ref: "#/components/parameters/LfRunId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SparseLabelMatrix"
  /v1/probabilistic-labels:
    get:
      tags: [probabilisticLabels]
      operationId: listProbabilisticLabels
      parameters:
        - in: query
          name: tag_id
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProbabilisticLabel"
    post:
      tags: [probabilisticLabels]
      operationId: upsertProbabilisticLabel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProbabilisticLabelUpsert"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProbabilisticLabel"
...
components:
  parameters:
...
    LfRunId:
      name: run_id
      in: path
      required: true
      schema: { type: string }
  schemas:
...
    LfVote:
      type: integer
      enum: [-1, 0, 1]
      description: 1 positive for tag, -1 negative, 0 abstain
    LfRunStatus:
      type: string
      enum: [pending, running, completed, failed]
    LfRunCreate:
      type: object
      required: [tag_id, labeling_function_ids]
      properties:
        tag_id: { type: string }
        labeling_function_ids:
          type: array
          items: { type: string }
          minItems: 1
    LfRun:
      type: object
      required: [id, tag_id, labeling_function_ids, status, created_at]
      properties:
        id: { type: string }
        tag_id: { type: string }
        labeling_function_ids:
          type: array
          items: { type: string }
        status:
          $ref: "#/components/schemas/LfRunStatus"
        error: { type: string, nullable: true }
        documents_scanned: { type: integer }
        votes_written: { type: integer }
        created_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time, nullable: true }
    SparseLabelMatrix:
      type: object
      required: [run_id, document_ids, labeling_function_ids, entries]
      properties:
        run_id: { type: string }
        document_ids:
          type: array
          items: { type: string }
        labeling_function_ids:
          type: array
          items: { type: string }
        entries:
          type: array
          items:
            $ref: "#/components/schemas/SparseLabelMatrixEntry"
    SparseLabelMatrixEntry:
      type: object
      required: [d, l, v]
      properties:
        d: { type: integer, description: Index into document_ids }
        l: { type: integer, description: Index into labeling_function_ids }
        v:
          $ref: "#/components/schemas/LfVote"
    ProbabilisticLabel:
      type: object
      required: [document_id, tag_id, probability]
      properties:
        document_id: { type: string }
        tag_id: { type: string }
        probability: { type: number, minimum: 0, maximum: 1 }
        conflict_score: { type: number, nullable: true }
        entropy: { type: number, nullable: true }
        updated_at: { type: string, format: date-time }
    ProbabilisticLabelUpsert:
      type: object
      required: [document_id, tag_id, probability]
      properties:
        document_id: { type: string }
        tag_id: { type: string }
        probability: { type: number, minimum: 0, maximum: 1 }
        conflict_score: { type: number, nullable: true }
        entropy: { type: number, nullable: true }
...